+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 * Compatibility with vshard configuration if UUIDs are omitted (#407).
+* Compatibility with automatic master discovery in vshard (#409).
 
 ## [1.4.2] - 25-12-23